requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
torch
torchvision
numpy==1.26.4
scikit-learn==1.5.2
opencv-python==4.10.0.84
Pillow==10.4.0
python-multipart==0.0.6